Apply suggestions from review comments
